Make stop registry seed inserter methods callable without running whole seed script
