+++ test-db-manager/src/seedStopRegistry.ts
@@ -1,181 +1,17 @@
-/* eslint-disable no-console */
+import { seedOrganisations, seedStopAreas, seedStopPlaces } from './datasets';
 import {
-  StopAreaInput,
-  StopPlaceInput,
-  StopPlaceNetexRef,
-  seedOrganisations,
-  seedStopAreas,
-  seedStopPlaces,
-} from './datasets';
-import {
-  StopRegistryOrganisationInput,
-  StopRegistryVersionLessEntityRefInput,
-} from './generated/graphql';
-import { insertStopPlaceForScheduledStopPoint } from './graphql-helpers';
-import { hasuraApi } from './hasuraApi';
-import {
-  mapToInsertOrganisationMutation,
-  mapToInsertStopAreaMutation,
-} from './queries';
-import { mapToGetStopPointByLabelQuery } from './queries/routesAndLines';
-import {
-  GetStopPointByLabelResult,
-  InsertOrganisationResult,
-  InsertStopAreaResult,
-} from './types';
-
-const insertOrganisation = async (
-  organisation: StopRegistryOrganisationInput,
-) => {
-  try {
-    const returnValue = (await hasuraApi(
-      mapToInsertOrganisationMutation(organisation),
-    )) as InsertOrganisationResult;
-
-    if (returnValue.data === null) {
-      throw new Error('Null data returned from Tiamat');
-    }
-
-    return returnValue.data.stop_registry.mutateOrganisation[0];
-  } catch (error) {
-    console.error(
-      'An error occurred while inserting organisation!',
-      organisation.name,
-      error,
-    );
-    throw error;
-  }
-};
-
-const insertStopPlace = async (
-  { label, maintenance, stopPlace }: StopPlaceInput,
-  organisationIdsByName: Map<string, string>,
-): Promise<StopPlaceNetexRef> => {
-  // Find related scheduled stop point.
-  const stopPointResult = (await hasuraApi(
-    mapToGetStopPointByLabelQuery(label),
-  )) as GetStopPointByLabelResult;
-
-  const stopPoints =
-    stopPointResult?.data?.service_pattern_scheduled_stop_point;
-  const stopPoint = stopPoints && stopPoints[0];
-  const stopPointId = stopPoint?.scheduled_stop_point_id;
-
-  if (!stopPointId) {
-    throw new Error(
-      `Could not find scheduled stop point with label ${label}. Did you forget to import route DB dump?`,
-    );
-  }
-  if (stopPoints.length > 1) {
-    // It is possible that there are multiple stop points for one label, with different priorities and/or validity times.
-    // With current dump this does not happen, but if things change, we might need some more logic here...
-    console.warn(
-      `Found multiple stop points for label ${label}, using the first one...`,
-    );
-  }
-
-  try {
-    const stopPlaceRef = await insertStopPlaceForScheduledStopPoint({
-      scheduledStopPointId: stopPointId,
-      stopPlace,
-      maintenance,
-      organisationIdsByName,
-    });
-
-    if (stopPoint.stop_place_ref) {
-      console.warn(
-        `Stop point ${label} (${stopPointId}) already has a stop place with id ${stopPoint.stop_place_ref}. Overwrote with ${stopPlaceRef}.`,
-      );
-    }
-    return { netexId: stopPlaceRef, label };
-  } catch (error) {
-    console.error(
-      'An error occurred while inserting stop place!',
-      label,
-      stopPlace,
-      error,
-    );
-    throw error;
-  }
-};
-
-const insertStopArea = async (
-  stopArea: StopAreaInput,
-  stopPlaces: Map<string, string>,
-) => {
-  const area = {
-    ...stopArea.stopArea,
-    members: stopArea.memberLabels.map(
-      (label) =>
-        ({
-          ref: stopPlaces.get(label),
-        }) as StopRegistryVersionLessEntityRefInput,
-    ),
-  };
-
-  try {
-    const returnValue = (await hasuraApi(
-      mapToInsertStopAreaMutation(area),
-    )) as InsertStopAreaResult;
-    if (returnValue.data === null) {
-      throw new Error('Null data returned from Tiamat');
-    }
-  } catch (error) {
-    console.error(
-      'An error occurred while inserting stop area!',
-      stopArea.stopArea.name?.value,
-      error,
-    );
-    throw error;
-  }
-};
+  insertOrganisations,
+  insertStopAreas,
+  insertStopPlaces,
+} from './graphql-helpers';
 
 const seedStopRegistry = async () => {
-  const collectedOrganisationIds: Map<string, string> = new Map();
-  console.log('Inserting organisations...');
-  // Need to run these sequentially. Will get transaction errors if trying to do concurrently.
-  for (let index = 0; index < seedOrganisations.length; index++) {
-    const organisation = seedOrganisations[index];
-    console.log(
-      `Organisation ${organisation.name}: organisation insert starting...`,
-    );
-    // eslint-disable-next-line no-await-in-loop
-    const result = await insertOrganisation(organisation);
-    collectedOrganisationIds.set(result.name, result.id);
-    console.log(
-      `Organisation ${organisation.name}: organisation insert finished!`,
-    );
-  }
-  console.log(`Inserted ${seedOrganisations.length} organisations.`);
-  console.log('Seed organisation ids: ', collectedOrganisationIds);
-
-  const collectedStopIds: Map<string, string> = new Map();
-  console.log('Inserting stop places...');
-  for (let index = 0; index < seedStopPlaces.length; index++) {
-    const stopPoint = seedStopPlaces[index];
-    console.log(
-      `Stop point ${stopPoint?.label}: stop place insert starting...`,
-    );
-    // eslint-disable-next-line no-await-in-loop
-    const netexRef = await insertStopPlace(stopPoint, collectedOrganisationIds);
-    collectedStopIds.set(netexRef.label, netexRef.netexId);
-    console.log(`Stop point ${stopPoint?.label}: stop place insert finished!`);
-  }
-  console.log(`Inserted ${seedStopPlaces.length} stop places.`);
-
-  console.log('Inserting stop areas...');
-  for (let index = 0; index < seedStopAreas.length; index++) {
-    const stopArea = seedStopAreas[index];
-    console.log(
-      `Stop area ${stopArea?.stopArea.name?.value}: stop area insert starting...`,
-    );
-    // eslint-disable-next-line no-await-in-loop
-    await insertStopArea(stopArea, collectedStopIds);
-    console.log(
-      `Stop area ${stopArea?.stopArea.name?.value}: stop area insert finished!`,
-    );
-  }
-  console.log(`Inserted ${seedStopAreas.length} stop areas.`);
+  const collectedOrganisationIds = await insertOrganisations(seedOrganisations);
+  const collectedStopIds = await insertStopPlaces(
+    seedStopPlaces,
+    collectedOrganisationIds,
+  );
+  await insertStopAreas(seedStopAreas, collectedStopIds);
 };
 
 seedStopRegistry();
